carpool et userAccount dto

+++ src/main/java/covoit/repository/CarpoolRepository.java
@@ -11,14 +11,14 @@ import covoit.entities.Carpool;
  * Repository for Carpool entity.
  */
 @Repository
-public interface CarpoolRepository extends JpaRepository<Carpool, Long> {
+public interface CarpoolRepository extends JpaRepository<Carpool, Integer> {
     /**
      * Finds all carpools associated with a specific user account.
      *
      * @param userId the ID of the user account
      * @return a list of carpools associated with the user
      */
-    List<Carpool> findByUserAccounts_Id(Long userId);
+    List<Carpool> findByUserAccounts_Id(int userId);
     Carpool findById(int id);
 	List<Carpool> findAll();
 	Carpool findByName(String name);
